io_uring/net: fix overflow check in io_recvmsg_mshot_prep()
[ Upstream commit 8ede3db5061bb1fe28e2c9683329aafa89d2b1b4 ]

The "controllen" variable is type size_t (unsigned long).  Casting it
to int could lead to an integer underflow.

The check_add_overflow() function considers the type of the destination
which is type int.  If we add two positive values and the result cannot
fit in an integer then that's counted as an overflow.

However, if we cast "controllen" to an int and it turns negative, then
negative values *can* fit into an int type so there is no overflow.

Good: 100 + (unsigned long)-4 = 96  <-- overflow
 Bad: 100 + (int)-4 = 96 <-- no overflow

I deleted the cast of the sizeof() as well.  That's not a bug but the
cast is unnecessary.

Fixes: 9b0fc3c054ff ("io_uring: fix types in io_recvmsg_multishot_overflow")
Link: https://lore.kernel.org/r/138bd2e2-ede8-4bcc-aa7b-f3d9de167a37@moroto.mountain

+++ io_uring/net.c
@@ -433,10 +433,10 @@ static bool io_recvmsg_multishot_overflow(struct io_async_msghdr *iomsg)
 
 	if (iomsg->namelen < 0)
 		return true;
-	if (check_add_overflow((int)sizeof(struct io_uring_recvmsg_out),
+	if (check_add_overflow(sizeof(struct io_uring_recvmsg_out),
 			       iomsg->namelen, &hdr))
 		return true;
-	if (check_add_overflow(hdr, (int)iomsg->controllen, &hdr))
+	if (check_add_overflow(hdr, iomsg->controllen, &hdr))
 		return true;
 
 	return false;
